hotfix

+++ src/main/java/de/thundergames/gameplay/player/board/Utils.java
@@ -1,7 +1,7 @@
 /*
  * Copyright Notice for SwtPra10
  * Copyright (c) at ThunderGames | SwtPra10 2022
- * File created on 09.01.22, 16:05 by Carina Latest changes made by Carina on 09.01.22, 16:05 All contents of "Utils" are protected by copyright. The copyright law, unless expressly indicated otherwise, is
+ * File created on 17.01.22, 19:10 by Carina Latest changes made by Carina on 17.01.22, 19:10 All contents of "Utils" are protected by copyright. The copyright law, unless expressly indicated otherwise, is
  * at ThunderGames | SwtPra10. All rights reserved
  * Any type of duplication, distribution, rental, sale, award,
  * Public accessibility or other use
@@ -25,8 +25,7 @@ public class Utils {
   // This method was slightly modified.
   public static ArrayList<String> getRandomHSLAColor(int amount) {
     ArrayList<String> colors = new ArrayList<>(amount);
-    var huedelta = (int) (360 / amount);
-
+    var huedelta = 360 / amount;
     for (var i = 0; i < amount; i++) {
       var hue = i * huedelta;
       colors.add("hsla(" + hue + ", 100%, 50%, 1)");
